+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Changed**
 
-- The [openLinks](./mw/README.md#openlinks) extension for a Monaco editor now also supports MediaWiki-style external links
+- The [openLinks](./mw/README.md#openlinks) extension for a Monaco editor now also supports the `cite` HTML attribute and MediaWiki-style external links
 
 ## 2.18.7